xcpmd: Use kernel-provided SMBIOS/DMI info
The kernel already parses SMBIOS/DMI structures and exports their value
through sysfs.  Just read the values from there.  If you try to lseek to
get the size, it returns a bogus value of 4096.  That is why the read is
structured to keep reading until EOF.

This eliminates parsing code from xcpmd.  It also removes the need to
mmap /dev/mem.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>